rename: 전역에서 사용하는 framer-motion contants 위치 변경
