+++ src/features/churchInfo/ui/ChurchHistory.tsx
@@ -1,6 +1,6 @@
 import { motion } from "framer-motion";
-import { fadeInUp } from "@/features/churchInfo/constants/FadeInUp";
-import { stagger } from "@/features/churchInfo/constants/Stagger";
+import { fadeInUp } from "@/shared/constants/FadeInUp";
+import { stagger } from "@/shared/constants/Stagger";
 import { ChurchHistoryData } from "@/features/churchInfo/constants/ChurchHistoryData";
 
 export default function ChurchHistory() {
