+++ src/features/churchInfo/ui/ChurchStatistics.tsx
@@ -1,6 +1,6 @@
 import { motion } from "framer-motion";
-import { fadeInUp } from "@/features/churchInfo/constants/FadeInUp";
-import { stagger } from "@/features/churchInfo/constants/Stagger";
+import { fadeInUp } from "@/shared/constants/FadeInUp";
+import { stagger } from "@/shared/constants/Stagger";
 
 export default function ChurchStatistics() {
   return (
