+++ src/shared/constants/FadeInUp.tsx
